Add optional flags for drat

+++ R/drat.R
@@ -1,7 +1,7 @@
 #' Build a package and commit to drat repo
 #'
 #' @name drat
-#' @note Updated 2021-01-19.
+#' @note Updated 2021-02-06.
 #'
 #' @return `logical(1)`.
 #'
@@ -13,23 +13,38 @@ NULL
 
 .drat <- function(
     pkgDirs,
-    repoDir = file.path("~", "monorepo", "drat")
+    repoDir = file.path("~", "monorepo", "drat"),
+    check = TRUE,
+    pkgdown = TRUE,
+    deploy = TRUE
 ) {
     requireNamespaces(c("desc", "devtools"))
     assert(
         allAreDirs(pkgDirs),
-        isADir(repoDir)
+        isADir(repoDir),
+        isFlag(check),
+        isFlag(pkgdown),
+        isFlag(deploy)
     )
     pkgDirs <- realpath(pkgDirs)
-    wd <- getwd()
+    if (isTRUE(deploy)) {
+        wd <- getwd()
+    }
     lapply(
         X = pkgDirs,
         repoDir = repoDir,
-        FUN = function(pkgDir, repoDir) {
+        check = check,
+        pkgdown = pkgdown,
+        deploy = deploy,
+        FUN = function(pkgDir, repoDir, check, pkgdown, deploy) {
             descFile <- file.path(pkgDir, "DESCRIPTION")
             assert(isAFile(descFile))
-            check(path = pkgDir)
-            .pkgdownDeployToAWS(pkg = pkgDir)
+            if (isTRUE(check)) {
+                check(path = pkgDir)
+            }
+            if (isTRUE(pkgdown)) {
+                .pkgdownDeployToAWS(pkg = pkgDir)
+            }
             tarball <- devtools::build(pkg = pkgDir)
             assert(isAFile(tarball))
             drat::insertPackage(
@@ -38,41 +53,48 @@ NULL
                 action = "archive"
             )
             invisible(file.remove(tarball))
-            setwd(pkgDir)
-            name <- desc::desc_get_field(key = "Package", file = descFile)
-            version <- as.character(desc::desc_get_version(file = descFile))
-            today <- Sys.Date()
-            shell(command = "git", args = c("fetch", "--force", "--tags"))
-            shell(
-                command = "git",
-                args = c(
-                    "tag", "--force",
-                    "-a", paste0("'v", version, "'"),
-                    "-m", paste0("'", name, " v", version, " (", today, ")'")
+            if (isTRUE(deploy)) {
+                setwd(pkgDir)
+                name <- desc::desc_get_field(key = "Package", file = descFile)
+                version <- as.character(desc::desc_get_version(file = descFile))
+                today <- Sys.Date()
+                shell(command = "git", args = c("fetch", "--force", "--tags"))
+                shell(
+                    command = "git",
+                    args = c(
+                        "tag",
+                        "--force",
+                        "-a",
+                        paste0("'v", version, "'"),
+                        "-m",
+                        paste0("'", name, " v", version, " (", today, ")'")
+                    )
                 )
-            )
-            shell(command = "git", args = c("push", "--force", "--tags"))
-            setwd(repoDir)
-            shell(command = "git", args = c("checkout", "master"))
-            shell(command = "git", args = c("fetch", "--all"))
-            shell(command = "git", args = "merge")
-            shell(command = "git", args = c("add", "./"))
-            shell(
-                command = "git",
-                args = c(
-                    "commit", "-m", paste0("'Add ", basename(tarball), ".'")
+                shell(command = "git", args = c("push", "--force", "--tags"))
+                setwd(repoDir)
+                shell(command = "git", args = c("checkout", "master"))
+                shell(command = "git", args = c("fetch", "--all"))
+                shell(command = "git", args = "merge")
+                shell(command = "git", args = c("add", "./"))
+                shell(
+                    command = "git",
+                    args = c(
+                        "commit", "-m", paste0("'Add ", basename(tarball), ".'")
+                    )
                 )
-            )
-            shell(command = "git", args = "push")
-            setwd(wd)
+                shell(command = "git", args = "push")
+                setwd(wd)
+            }
             message(sprintf("Successfully added '%s'.", basename(tarball)))
             TRUE
         }
     )
-    setwd(repoDir)
-    assert(isAFile("update"))
-    shell(command = "./update")
-    setwd(wd)
+    if (isTRUE(deploy)) {
+        setwd(repoDir)
+        assert(isAFile("update"))
+        shell(command = "./update")
+        setwd(wd)
+    }
     invisible(TRUE)
 }
 
@@ -129,10 +151,22 @@ NULL
 drat <- function() {
     parse <- parseArgs(
         optional = "repo-dir",
+        flags = c(
+            "no-check",
+            "no-deploy",
+            "no-pkgdown"
+        ),
         positional = TRUE
     )
-    args <- list("pkgDirs" = parse[["positional"]])
     optional <- parse[["optional"]]
+    flags <- parse[["flags"]]
+    positional <- parse[["positional"]]
+    args <- list(
+        "pkgDirs" = positional,
+        "check" = !isSubset("no-check", flags),
+        "deploy" = !isSubset("no-deploy", flags),
+        "pkgdown" = !isSubset("no-pkgdown", flags),
+    )
     if (!is.null(optional)) {
         if (isSubset("repo-dir", names(optional))) {
             args[["repoDir"]] <- optional[["repo-dir"]]
